implement single_ip
it was necessary to override the AnyUserIfPublic Generator to allow
only if the access_right is set to open, otherwise it would always be
allowed.

the single_ip need is now added to the identity with the
@identity_loaded.connect hook.

RECORDS_PERMISSIONS_RECORD_POLICY seams still not to work, there we
need another solution for the moment.

RECORDS_UI_ENDPOINTS could override the permissions for the record
landing page.